requirements.txt
absl-py==1.0.0
asgiref==3.4.1
astunparse==1.6.3
boto3==1.20.26
botocore==1.23.26
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
cycler==0.11.0
defusedxml==0.7.1
Django==3.2.9
django-storages==1.12.3
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flatbuffers==2.0
fonttools==4.28.3
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.8.2
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.0
nltk==3.6.5
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
s3transfer==0.5.0
setuptools-scm==6.3.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
tomli==1.2.2
tqdm==4.62.3
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2
whitenoise==5.3.0
wrapt==1.13.3
zipp==3.6.0
